% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wechat.R
\name{WechatMsg}
\alias{WechatMsg}
\title{WechatHelper}
\usage{
WechatMsg(userid, msg, status = NULL, name = NULL, number = NULL)
}
\arguments{
\item{userid}{请关注公众号"生信通知助手"，回复id，扫描二维码获取userid}

\item{msg}{需要发生的通知内容}

\item{status}{当前数据状态}

\item{name}{项目名称}

\item{number}{通知编号}
}
\value{
none
}
\description{
WechatHelper
}
