% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wechat.R
\name{WechatRobot}
\alias{WechatRobot}
\title{WechatRobot}
\usage{
WechatRobot(
  key,
  msg = "",
  qyapi = "https://qyapi.weixin.qq.com/cgi-bin/webhook/send?key=",
  msgtype = "text"
)
}
\arguments{
\item{key}{The key for the WeChat Work robot.}

\item{msg}{Text message.}

\item{qyapi}{WeChat Work API.}

\item{msgtype}{Type of text message being sent.}
}
\value{
response of the request
}
\description{
WechatRobot
}
\examples{
WechatRobot(key = "79bdbb23-7f7e-4fb6-b669-97f7e744755e", msg = "test---SaveRds Success!")
}
