% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wechat.R
\name{WechatRobot_R}
\alias{WechatRobot_R}
\title{WechatRobot_R}
\usage{
WechatRobot_R(
  key,
  msg = "",
  qyapi = "https://qyapi.weixin.qq.com/cgi-bin/webhook/send?key=",
  msgtype = "text"
)
}
\arguments{
\item{msgtype}{}
}
\description{
WechatRobot_R
}
