% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wechat.R
\name{WechatMsg}
\alias{WechatMsg}
\title{WechatHelper}
\usage{
WechatMsg(userid, msg, status = NULL, name = NULL, number = NULL)
}
\arguments{
\item{userid}{Please follow the official WeChat account "shengxintongzhizhushou", reply with id, and scan the QR code to obtain the userid. View "http://bj.s1f.ren/gzh/home.html" and "https://mp.weixin.qq.com/s/tiedh3FWTA4hOhut5Useew" for more information.}

\item{msg}{Content of the notification to be sent. Limit to 20 characters.}

\item{status}{Current data status. Limit to 5 Chinese characters.}

\item{name}{Project name. Limit to 5 characters.}

\item{number}{Notification number.}
}
\value{
response of the request
}
\description{
WechatHelper
}
\examples{
WechatMsg(userid = "123", msg = "test---SaveRds Success!")
}
